Fix some compiler warns (adchub and plg man)

+++ dcpp/AdcHub.cpp
@@ -571,7 +571,7 @@ void AdcHub::handle(AdcCommand::GET, AdcCommand& c) noexcept {
 		size_t n = ShareManager::getInstance()->getSharedFiles();
 
 		// When h >= 32, m can't go above 2^h anyway since it's stored in a size_t.
-        if(m > (5 * Util::roundUp((int64_t)(n * k / log(2.)), (int64_t)64)) || (h < 32 && m > static_cast<size_t>(1U << h))) {
+        if(m > (5 * (size_t)Util::roundUp((int64_t)(n * k / log(2.)), (int64_t)64)) || (h < 32 && m > static_cast<size_t>(1U << h))) {
 			send(AdcCommand(AdcCommand::SEV_FATAL, AdcCommand::ERROR_TRANSFER_GENERIC,
 				"Unsupported m", AdcCommand::TYPE_HUB));
 			return;
